modify pytest
