+++ proj/example.py
@@ -3,33 +3,8 @@ import numpy as np
 import pytest
 import cv2
 
-def trans2image(labels, img, k):
-
-    h, w = img.shape[0], img.shape[1]
-    output = np.zeros((h, w, 3))
-    for cls in range(k):
-        idxs = np.where(labels == cls)[0]
-
-        mean_color = 0.0
-        for idx in idxs:
-            y = idx // w
-            x = idx % w
-            #print(idx)
-            mean_color += (img[y][x] / len(idxs))
-
-        #print(mean_color.shape)
-        for idx in idxs:
-            y = idx // w
-            x = idx % w
-            output[y, x, :] = mean_color
-    
-    return output.astype(np.uint8)
-    
-
-
-
-def test_unit():
-    k = 2
+def cluster():
+    k = 3
     gamma_c=0.6
     gamma_s=0.4
     max_iter=20
@@ -46,28 +21,13 @@ def test_unit():
     assert c.nthreads == nthreads
     assert c.thresh == thresh
 
-    # print(c.k_cluster)
-    # print(c.gamma_c)
-    # print(c.gamma_s)
-    # print(c.max_iter)
-    # print(c.nthreads)
-    # print(c.thresh)
+    
 
+    print("------------------Do Cluster-----------------")
+    c.predict("./image1.png")
+    c.savefig("image1_result.png")
 
-    #print("---------------------------------------")
-
-    img = cv2.imread("./test.png")
-    c.predict(img)
-    #c.savefig("image1_result.png")
-
-    result = c.get_results()
-    assert result.shape[0] == img.shape[0] * img.shape[1]
-
-    #print(result)
-    output = trans2image(result, img, k)
-    #print((output - img).sum())
-    assert (output - img).sum() == 0.0
 
 
 if __name__ == "__main__":
-    test_unit()
+    cluster()
